requirements.txt
python-dotenv==1.0.1
requests==2.32.3
tqdm==4.66.5
faster-whisper==1.0.3
numpy==1.26.4
pyttsx3==2.90
edge-tts==6.1.16
gTTS==2.5.4
Flask==3.1.0